pyproject.toml
[project]
name = "indi-engine"
version = "0.1.0"
description = "Middleware layer between INDI server and clients"
requires-python = ">=3.13"
dependencies = [
    "pyyaml",
]


[project.scripts]
indi-engine = "indi_engine.main:main"

[dependency-groups]
dev = [
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["indi_engine"]
